% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R, R/read_bed.R
\name{check_files}
\alias{check_files}
\title{check_file, check is file path are correct}
\usage{
check_files(files)

check_files(files)
}
\arguments{
\item{files}{a vector of bed files path}
}
\value{
save a list of preprocessed bed files

save a list of preprocessed bed files
}
\description{
check_file, check is file path are correct

check_file, check is file path are correct
}
