% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{compute_hawkes_histogram}
\alias{compute_hawkes_histogram}
\title{getparam_overlap}
\usage{
compute_hawkes_histogram(
  K,
  delta,
  kernel,
  processes,
  path_data,
  path_results,
  ofile_prefix
)
}
\arguments{
\item{delta}{}

\item{kernel}{}

\item{ofile_prefix}{}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
