% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{compute_hawkes_overlap}
\alias{compute_hawkes_overlap}
\title{My median}
\usage{
compute_hawkes_overlap(
  K,
  delta,
  kernel,
  processes,
  path_data,
  path_results,
  ofile_prefix
)
}
\arguments{
\item{x}{Vector of Numeric values}
}
\value{
Median of vector x
}
\description{
My median
}
