% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{df2bedr}
\alias{df2bedr}
\title{getparam_overlap}
\usage{
df2bedr(df, check.valid = FALSE)
}
\arguments{
\item{df}{}

\item{check.valid}{(default: FALSE)}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
