% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{get_regions_from_bed}
\alias{get_regions_from_bed}
\title{getparam_overlap}
\usage{
get_regions_from_bed(dd, maxlag, check.valid = FALSE)
}
\arguments{
\item{dd}{}

\item{check.valid}{(default: FALSE)}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
