% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{getparam}
\alias{getparam}
\title{getparam_overlap}
\usage{
getparam(ofile, normalize = FALSE)
}
\arguments{
\item{ofile}{}

\item{normalize}{(default: FALSE)}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
