% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{getparam_overlap}
\alias{getparam_overlap}
\title{getparam_overlap}
\usage{
getparam_overlap(M, K, ofile, normalize = FALSE)
}
\arguments{
\item{M}{}

\item{normalize}{}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
