% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{hw}
\alias{hw}
\title{getparam_overlap}
\usage{
hw(a, K, delta, eta_m, eta_ell)
}
\arguments{
\item{eta_m}{}

\item{eta_ell}{}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
