% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{hw convolution of the h interaction function with the kernel of the 2
processus}
\usage{
plot_histogram(data, K = 10, delta = 10000)
}
\arguments{
\item{data}{results of compute_hawkes_histogram}

\item{K}{(default: 10) size of the histogram bins}

\item{delta}{(default: 1e4) max range to consider}

\item{width}{a vector of beds interval width (in the same orde as the data columns)}
}
\value{
tibble
}
\description{
hw convolution of the h interaction function with the kernel of the 2
processus
}
