% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_overlap.R
\name{results_hawkes_overlap}
\alias{results_hawkes_overlap}
\title{getparam_overlap}
\usage{
results_hawkes_overlap(
  K,
  delta,
  kernel,
  processes,
  path_data,
  path_results,
  ofile_prefix
)
}
\arguments{
\item{delta}{}

\item{kernel}{}

\item{ofile_prefix}{}
}
\value{
Median of vector x
}
\description{
getparam_overlap
}
