% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple_convolution.R
\name{triple_convolution}
\alias{triple_convolution}
\title{the uniform kernel of width width}
\usage{
triple_convolution(position, x, width, vs_width, delta = 10000)
}
\arguments{
\item{position}{position on which to convolve}

\item{x}{interaction function parameter}

\item{width}{width of the uniform kernel for the first process}

\item{vs_width}{width of the uniform kernel for the second process}

\item{delta}{(default: 1e4) max range to consider}
}
\value{
Median of vector x
}
\description{
the uniform kernel of width width
}
