% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.R, R/uniform_kernel.R
\name{uniform_kernel}
\alias{uniform_kernel}
\title{the uniform kernel of width width}
\usage{
uniform_kernel(x, width)

uniform_kernel(x, width)
}
\arguments{
\item{x}{vector of data}

\item{width}{width of the uniform kernel}
}
\value{
Median of vector x

Median of vector x
}
\description{
the uniform kernel of width width

the uniform kernel of width width
}
