% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hawkes_histogram.R, R/read_bed.R
\name{compute_hawkes_histogram}
\alias{compute_hawkes_histogram}
\title{compute_hawkes_histogram}
\usage{
compute_hawkes_histogram(
  files,
  names = files,
  K = 10,
  delta = 10000,
  kernel = "heterogeneous_interval"
)

compute_hawkes_histogram(
  files,
  names = files,
  K = 10,
  delta = 10000,
  kernel = "heterogeneous_interval"
)
}
\arguments{
\item{files}{a vector of bed files path}

\item{names}{(default: files) names of the samples}

\item{K}{(default: 10) size of the histogram bins}

\item{delta}{(default: 1e4) max range to consider}

\item{kernel}{(default: "heterogeneous_interval") kernel for the computation}
}
\value{
save a results file

save a results file
}
\description{
compute_hawkes_histogram

compute_hawkes_histogram
}
