% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getparam.R, R/read_bed.R
\name{getparam}
\alias{getparam}
\title{getparam}
\usage{
getparam(data, K = 10, delta = 10000)

getparam(data, K = 10, delta = 10000)
}
\arguments{
\item{data}{a matrix of parameters}

\item{K}{(default: 10) size of the histogram bins}

\item{delta}{(default: 1e4) max range to consider}
}
\value{
Median of vector x

Median of vector x
}
\description{
getparam

getparam format results from the output of compute_hawkes_histogram
}
